add_mail can now store mails in mails.db and add the mail metadata (from, to, email hash) in the address db
user_id returns the uid of the user from the keys db. (the primary key 
of the row is the user_id)

add_user inserts the keys of the client into the keys db and also 
timestamps his login. It then creates another table in the addresses db 
to store the mailbox logs